symbol: Load dynamic symbols for bind-now binaries
When "bind-now" feature was enabled, linker resolves all dynamic symbols
at load time and the resolver function will not be called at runtime.
This also removes relocation entries for the PLT functions so uftrace
cannot know location for hooking.

In this case, it needs to load symbols by reading instructions in
.plt.got section and find matching entry in dynamic symbol table.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>