arch/x86: Implement mcount_arch_undo_bindnow()
When "bind-now" is now, the binary doesn't have PLT (other than PLT0)
for the external functions.  However, in order to trace those functions,
it needs to have PLT so that the plt_hooker can be called.

This patch rebuilt the PLT like dynamic tracing does, but it allocates
the trampoline in a separate memory region since it can be large.  A
function in PLT will have a 16 byte entry for PUSH, JMP instructions and
padding.  As symbol index can be different to the index in GOT, it
calculate GOT index using jump offset in the instructions.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>