report: Consolidate unpaired entries with a dummy
From now on, perf report --diff shows entries which does not have
matching entry (pair) in the different data.  But it sometimes annoying
that small differences in one data can hide more important (bigger) diff
output.  So it'd be better combining them and show together - sorted by
the difference in the end.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>